# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ai/translate": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "A011:Translate",
        "description": "テキスト翻訳エンドポイント（非推奨）\n\nArgs:\n    body (TranslateTextRequest): 翻訳リクエストボディ\n\nReturns:\n    TranslatedText: 翻訳結果（現在は未実装）\n\nNote:\n    このエンドポイントは非推奨です。",
        "operationId": "A011",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslatedText"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/api/ai/chat/start": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "A012:Start Chat",
        "description": "ユーザのリクエストに応じて初期コレクションを作成する\n\nユーザーのメッセージに基づいて、関連するドキュメントを検索し、\nそれらのドキュメントを含む新しいコレクションを作成します。\n\nArgs:\n    body (ChatRequestBase): リクエストボディ。メッセージデータが含まれる\n    user (User): 認証済みユーザー情報\n    service (AIService): AIサービス\n\nReturns:\n    CollectionDetail: 作成されたコレクションの詳細情報\n\nRaises:\n    HTTPException: 以下の場合にHTTPExceptionが発生します：\n        - 500: コレクション作成に失敗した場合\n        - 500: 内部サーバーエラーが発生した場合",
        "operationId": "A012",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ai/chat/completions": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "A013:Chat Completions",
        "description": "ユーザーのリクエストに応じたチャットをストリーミングで返す\n\n指定されたコレクションと会話履歴を基に、AIエージェントが生成した\nレスポンスをServer-Sent Eventsでストリーミング配信します。\n新規の場合は会話を作成し、既存の場合は会話を継続します。\n\nArgs:\n    body (ChatCompletionsRequest): チャットのリクエスト\n    user (User): 認証済みユーザー情報\n    service (AIService): AIサービスインスタンス\n\nReturns:\n    StreamingResponse: 以下の情報をJSONパース可能な文字列としてストリームします：\n        - Message: メッセージオブジェクト\n        - Artifact: アーティファクトオブジェクト\n        - public_conversation_id: 会話ID文字列\n\nRaises:\n    HTTPException: 以下の場合にHTTPExceptionが発生します：\n        - 404: 指定されたコレクションが存在しない場合\n        - 404: 指定された会話履歴が存在しない場合\n        - 403: ユーザーがコレクションへのアクセス権限を持たない場合\n        - 500: チャット処理に失敗した場合\n        - 500: 内部サーバーエラーが発生した場合",
        "operationId": "A013",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamChat"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/StreamChat"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_authorized": {
                    "value": {
                      "detail": "Not authorized to use this collection"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_found": {
                    "value": {
                      "detail": "Collection not found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Artifact": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "成果物のID",
            "examples": [
              "3a92c5d1-6f3b-4b9a-bc2c-9f7d6e4f8e7a"
            ]
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "成果物のバージョン",
            "examples": [
              1
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "成果物のタイトル",
            "examples": [
              "CANドライバの初期化手順"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "成果物の内容",
            "examples": [
              "CANドライバの初期化手順は、通常以下のようなステップで..."
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "title",
          "content"
        ],
        "title": "Artifact"
      },
      "ArtifactBase": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "成果物のID",
            "examples": [
              "3a92c5d1-6f3b-4b9a-bc2c-9f7d6e4f8e7a"
            ]
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "成果物のバージョン",
            "examples": [
              1
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "version"
        ],
        "title": "ArtifactBase"
      },
...
      "ChatCompletionsRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "ユーザーの入力したAIへの要求内容",
            "examples": [
              "CANドライバの初期化手順はどのように定義されていますか？"
            ]
          },
          "public_collection_id": {
            "type": "string",
            "title": "Public Collection Id",
            "description": "公開コレクションID",
            "examples": [
              "V1StGXR8_Z5"
            ]
          },
          "public_conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Conversation Id",
            "description": "会話履歴の公開ID",
            "examples": [
              "Zf2k6A6R5h2"
            ]
          }
        },
        "type": "object",
        "required": [
          "message",
          "public_collection_id"
        ],
        "title": "ChatCompletionsRequest",
        "description": "チャット完了のリクエストモデル"
      },
      "ChatRequestBase": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "ユーザーの入力したAIへの要求内容",
            "examples": [
              "CANドライバの初期化手順はどのように定義されていますか？"
            ]
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequestBase",
        "description": "チャットリクエストの基底クラス"
      },
...
      "CollectionDetail": {
        "properties": {
          "public_collection_id": {
            "type": "string",
            "title": "Public Collection Id",
            "description": "公開コレクションID",
            "examples": [
              "V1StGXR8_Z5"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "コレクション名",
            "examples": [
              "CAN仕様"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "コレクションの作成日時",
            "examples": [
              "2023-05-16T14:30:45.123456+09:00"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "コレクションの更新日時",
            "examples": [
              "2023-05-16T14:30:45.123456+09:00"
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentBase"
            },
            "type": "array",
            "title": "Documents",
            "description": "コレクションに含まれるドキュメントのリスト",
            "examples": [
              [
                {
                  "file_type": "application/pdf",
                  "genre": "SWS",
                  "id": "260ea4dd-36a7-401f-b2a1-d6a28d97f140",
                  "path": "/app/public/documents/R22-11/AUTOSAR_SWS_CANDriver.pdf",
                  "release": "R22-11",
                  "subject": "AUTOSAR",
                  "title": "Specification of CAN Driver"
                }
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "public_collection_id",
          "name",
          "created_at",
          "updated_at",
          "documents"
        ],
        "title": "CollectionDetail"
      },
      "Context": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "ドキュメントのタイトル",
            "examples": [
              "CAN Driver Specification"
            ]
          },
          "chunk": {
            "type": "string",
            "title": "Chunk",
            "description": "回答に使用された根拠となる情報",
            "examples": [
              "...このドキュメントではCAN Driverの仕様を記述..."
            ]
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "ドキュメントのパス",
            "examples": [
              "/app/public/documents/R22-11/AUTOSAR_SWS_CANDriver.pdff"
            ]
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "`chunk`が含まれるページ番号",
            "examples": [
              12
            ]
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "ドキュメントのMIMEタイプ",
            "examples": [
              "application/pdf"
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "chunk",
          "path",
          "page",
          "file_type"
        ],
        "title": "Context"
      },
...
      "DocumentBase": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ドキュメントID",
            "examples": [
              "260ea4dd-36a7-401f-b2a1-d6a28d97f140"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "ドキュメントのタイトル",
            "examples": [
              "Specification of CAN Driver"
            ]
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "ドキュメントのパス",
            "examples": [
              "/app/public/documents/R22-11/AUTOSAR_SWS_CANDriver.pdf"
            ]
          },
          "subject": {
            "type": "string",
            "enum": [
              "AUTOSAR",
              "others"
            ],
            "title": "Subject",
            "description": "ドキュメントの種類",
            "examples": [
              "AUTOSAR"
            ]
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre",
            "description": "AUTOSARドキュメントのワーキンググループ",
            "examples": [
              "SWS"
            ]
          },
          "release": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release",
            "description": "AUTOSARドキュメントのリリースバージョン",
            "examples": [
              "R22-11"
            ]
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "ドキュメントのMIMEタイプ",
            "examples": [
              "application/pdf"
            ]
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "ドキュメントの要約",
            "default": "",
            "examples": [
              "This document describes the CAN driver specification..."
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "path",
          "subject",
          "file_type"
        ],
        "title": "DocumentBase",
        "description": "ドキュメントの基本情報"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "description": "メッセージのロール(user/assistant)",
            "examples": [
              "assistant"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "メッセージの内容",
            "examples": [
              "CANドライバの初期化手順は、通常以下のようなステップで..."
            ]
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata",
            "description": "メッセージに紐づくメタデータ"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "metadata"
        ],
        "title": "Message"
      },
      "MessageMetadata": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "description": "メタデータのバージョン",
            "default": "v1"
          },
          "contexts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Context"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexts",
            "description": "回答に使用したコンテキスト"
          },
          "recommended_documents": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DocumentBase"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Documents",
            "description": "レコメンドしたドキュメントの情報",
            "examples": [
              [
                {
                  "file_type": "application/pdf",
                  "genre": "SWS",
                  "id": "260ea4dd-36a7-401f-b2a1-d6a28d97f140",
                  "path": "/app/public/documents/R22-11/AUTOSAR_SWS_CANDriver.pdf",
                  "release": "R22-11",
                  "subject": "AUTOSAR",
                  "title": "Specification of CAN Driver"
                }
              ]
            ]
          },
          "generated_artifacts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ArtifactBase"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Artifacts",
            "description": "このメッセージと一緒に生成された成果物のIDとバージョン",
            "examples": [
              [
                {
                  "id": "123e4567-e78b-3fea-15j6d93o931",
                  "version": 1
                }
              ]
            ]
          }
        },
        "type": "object",
        "title": "MessageMetadata"
      },
...
      "StreamChat": {
        "properties": {
          "public_conversation_id": {
            "type": "string",
            "title": "Public Conversation Id",
            "description": "会話履歴の公開ID",
            "examples": [
              "Zf2k6A6R5h2"
            ]
          },
          "message": {
            "$ref": "#/components/schemas/Message",
            "description": "AIの返答内容"
          },
          "artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Artifact"
              },
              {
                "type": "null"
              }
            ],
            "description": "チャットに紐づく成果物"
          }
        },
        "type": "object",
        "required": [
          "public_conversation_id",
          "message"
        ],
        "title": "StreamChat",
        "description": "チャット完了時のストリーミングレスポンスモデル"
      },
      "TranslateTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "日本語に翻訳したいテキスト",
            "examples": [
              "The initialization procedure for the CAN driver usually consists of the following steps..."
            ]
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TranslateTextRequest",
        "description": "テキスト翻訳のリクエストモデル"
      },
      "TranslatedText": {
        "properties": {
          "result": {
            "type": "string",
            "title": "Result",
            "description": "翻訳結果（日本語）",
            "examples": [
              "CANドライバの初期化手順は、通常以下のようなステップで..."
            ]
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "TranslatedText",
        "description": "翻訳結果のレスポンスモデル"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}